% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp_2016_odds.R
\name{atp_odds}
\alias{atp_odds}
\title{2016 ATP Match Odds}
\format{A data frame with 2,626 rows and 10 variables}
\source{
\url{http://www.tennis-data.co.uk/alldata.php}
}
\description{
This dataset contains outcomes of matches on the ATP Tour in 2016. The variables of the dataset are:
}
\details{
\itemize{
  \item Location. A character venue of tournament
  \item Tournament. A character name of tounament (including sponsor if relevant)
  \item Date. A character indicating  the date of match 
  \item Series. A character of name of ATP tennis series 
  \item Court. A character indicating the type of court (Outdoor or Indoor)
  \item Surface. A character indicating the type of surface (Clay, Hard, Carpet or Grass)
  \item Round. A character indicating the round of match
  \item Best.of. A character with the maximum number of sets playable in match (3 or 5)
  \item Winner. A character (Last First Initial) name of the match winner
  \item Loser. A character (Last First Initial) name of the match loser
}
}

