% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_stats.R
\name{diff_stats}
\alias{diff_stats}
\title{Statistic Differences of ATP Matches}
\format{A data frame with 2,852 rows and 19 variables}
\description{
This dataset contains difference statistics for 2016 ATP matches. The variables of the dataset are:
}
\details{
\itemize{
  \item tourney_name. A character tournament name
  \item surface. A character description of the court surface (Carpet, Clay, Grass, or Hard)
  \item tourney_level. A character indicating the tournament level
  \item tourney_date. A Date object of the tournament date
  \item winner_name. A character of the winner's name
  \item loser_name. A character of the loser's name
  \item tourney_date. A Date object of the tournament date
  \item diff_ace_rate. A numeric that is the difference in rate of aces between player 1 and player2.
  \item diff_df_rate. A numeric that is the difference in rate of double faults between player 1 and player2.
  \item diff_first_serve_in_p. A numeric that is the difference in first serve proportion in between player 1 and player2.
  \item diff_first_serve_won_p. A numeric that is the difference in proportion of first service points won between player 1 and player2.
  \item diff_second_serve_won_p. A numeric that is the difference proportion of second service points won between player 1 and player2.
  \item diff_serve_won_p. A numeric that is the difference in proportion service points won player 1 and player2.
  \item diff_first_return_won_p. A numeric that is the difference in proportion of first return points won between player 1 and player2.
  \item diff_second_return_won_p. A numeric that is the difference in proportion of second return points won between player 1 and player2.
  \item diff_return_won_p. A numeric that is the difference in proportion of return points won between player 1 and player2.
  \item diff_total_won_p. A numeric that is the difference in proportion of total points won between player 1 and player2.
  \item diff_bp. A numeric that is the difference in break points won between player 1 and player2.
  \item diff_bp_opp. A numeric that is the difference in break point opportunities between player 1 and player2.
  \item winner. An integer indicating whether player1 was the winner = 1 or loser = 0 of the match.
}
}

